students/Y2332/Timofeev_Mikhail/lab1/GIBDD_proj/requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.1.7
django-cors-headers==3.7.0
django-rest-swagger==2.2.0
django-summernote==0.8.11.6
django-templated-mail==1.1.1
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-jsonapi==4.2.0
djoser==2.1.0
Faker==8.2.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.1.0
openapi-codec==1.3.2
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
simplejson==3.17.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.0.2
sqlparse==0.4.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.26.4